refactor: user controller split
